Apply pr suggestions
