+++ packages/airnode-utilities/src/constants.ts
@@ -1,0 +1,6 @@
+// The default amount of time before a "retryable" promise is timed out and retried
+export const DEFAULT_RETRY_TIMEOUT_MS = 5_000;
+// The Priority Fee in Wei
+export const PRIORITY_FEE_IN_WEI = 3120000000;
+// The Base Fee to Max Fee multiplier
+export const BASE_FEE_MULTIPLIER = 2;
