+++ packages/airnode-utilities/src/evm/index.ts
@@ -1,0 +1,1 @@
+export * from './gas-prices';
