+++ packages/airnode-utilities/src/index.ts
@@ -4,4 +4,4 @@ export * from './date';
 export * from './promises';
 export * from './strings';
 export * from './logging';
-export * from './gas-prices';
+export * from './evm';
